+++ package.json
@@ -9,6 +9,7 @@
     "@types/node": "^14.6.0",
     "@types/react": "^16.9.46",
     "@types/react-dom": "^16.9.8",
+    "@types/react-modal": "^3.10.6",
     "@types/react-router-dom": "^5.1.5",
     "@types/styled-components": "^5.1.2",
     "axios": "^0.19.2",
